Created dynamic hook for image transition for Matrix and Gallery (not pushed)
