+++ src/Hooks.js
@@ -10,3 +10,22 @@ export const useFetch = (url, initialValue) => {
 
   return result;
 }
+
+export const useDynamicTransition = ({ increment, delay, length }) => {
+  const [index, setIndex] = useState(0)
+
+  useEffect(() => {
+    const interval = setInterval(
+      () => {
+        setIndex(storedIndex => {
+          return (storedIndex+1)%length
+        })
+      },
+      delay
+    )
+
+    return () => clearInterval(interval)
+  }, [delay, increment])
+
+  return index
+}
